+++ src/pages/Root.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { GlobalShell, ThemeButton } from '../components/common';
-import { Cards, Board } from '../components/rootPage';
+import { Banner, Board } from '../components/rootPage';
 
 const Root = () => {
   console.log('Root');
@@ -8,9 +8,8 @@ const Root = () => {
   return (
     <>
       <GlobalShell>
-        {/* <Banner /> */}
+        <Banner />
         <Board />
-        <Cards />
       </GlobalShell>
       <ThemeButton />
     </>
